+++ library/src/main/java/me/denley/preferencebinder/BindPref.java
@@ -3,6 +3,8 @@ package me.denley.preferencebinder;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
+import me.denley.preferencebinder.internal.WidgetBindingType;
+
 import static java.lang.annotation.ElementType.FIELD;
 import static java.lang.annotation.ElementType.METHOD;
 import static java.lang.annotation.RetentionPolicy.CLASS;
@@ -21,5 +23,8 @@ public @interface BindPref {
 
     /** Whether or not to update this field or call this method again when the preference value changes */
     boolean listen() default true;
+
+    /** The method to use to bind to a widget */
+    WidgetBindingType bindTo() default WidgetBindingType.ASSIGN;
 
 }